core/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
requests==2.32.3
PyJWT==2.9.0